samples: updated ProcessDocument to v1
