+++ samples/snippets/src/main/java/documentai/v1/QuickStart.java
@@ -51,8 +51,8 @@ public class QuickStart {
       // The full resource name of the processor, e.g.:
       // projects/project-id/locations/location/processor/processor-id
       // You must create new processors in the Cloud Console first
-      String name =
-          String.format("projects/%s/locations/%s/processors/%s", projectId, location, processorId);
+      String name = String.format("projects/%s/locations/%s/processors/%s", 
+              projectId, location, processorId);
 
       // Read the file.
       byte[] imageFileData = Files.readAllBytes(Paths.get(filePath));
@@ -60,12 +60,16 @@ public class QuickStart {
       // Convert the image data to a Buffer and base64 encode it.
       ByteString content = ByteString.copyFrom(imageFileData);
 
-      RawDocument document =
-          RawDocument.newBuilder().setContent(content).setMimeType("application/pdf").build();
+      RawDocument document = RawDocument.newBuilder()
+              .setContent(content)
+              .setMimeType("application/pdf")
+              .build();
 
       // Configure the process request.
-      ProcessRequest request =
-          ProcessRequest.newBuilder().setName(name).setRawDocument(document).build();
+      ProcessRequest request = ProcessRequest.newBuilder()
+              .setName(name)
+              .setRawDocument(document)
+              .build();
 
       // Recognizes text entities in the PDF document
       ProcessResponse result = client.processDocument(request);
